chore: re-organize functions
