+++ docs/_static/js/utils.js
@@ -16,15 +16,6 @@ const getModeIconSrc = (mode) => {
   return CLASSIC_ICON_PATH;
 }
 
-const addFooterNote = () => {
-  const contentInfo = document.querySelector("div[role=contentinfo]");
-  const footerNote = document.createElement("p");
-  footerNote.classList.add("footer-note");
-  footerNote.innerHTML =
-    'Customized with ❤️ by the <a href="https://ethereum.org/" target="_blank">ethereum.org</a> team.';
-  contentInfo.parentNode.insertBefore(footerNote, contentInfo.nextSibling);
-};
-
 const updateActiveNavLink = () => {
   const navLinks = document.querySelectorAll(".unified-header .nav-link");
   navLinks.forEach((link) => {
@@ -87,31 +78,6 @@ const updateColorModeIcon = (button) => {
   appendSvg(getModeIconSrc(mode), button, COLOR_TOGGLE_ICON_CLASS);
 };
 
-const addColorModeButton = () => {
-  // Create a new button element
-  const colorModeButton = document.createElement("button");
-  colorModeButton.classList.add("color-toggle");
-  colorModeButton.setAttribute("type", "button");
-  colorModeButton.setAttribute("aria-label", "Toggle light dark mode");
-  colorModeButton.setAttribute("key", "color mode button");
-  colorModeButton.onclick = cycleColorMode;  
-  // Update the icon for this button according to the current mode
-  updateColorModeIcon(colorModeButton);
-
-  // Select the side nav search container
-  const sideNavSearch = document.querySelector('.wy-side-nav-search');
-
-  // Remove the anchor and div.version siblings before the input
-  const anchor = sideNavSearch.querySelector('a');
-  const versionDiv = sideNavSearch.querySelector('div.version');
-  if (anchor) sideNavSearch.removeChild(anchor);
-  if (versionDiv) sideNavSearch.removeChild(versionDiv);
-
-  // Insert the button before the search input
-  const searchInput = sideNavSearch.querySelector('[role="search"]');
-  sideNavSearch.insertBefore(colorModeButton, searchInput);
-}
-
 const cycleColorMode = () => {
   if (!COLOR_CHOICES.includes(mode)) return;
 
@@ -125,135 +91,12 @@ const cycleColorMode = () => {
   updateColorModeIcon(colorModeButton);
 }
 
-const preloadFonts = () => {
-  const fonts = [
-    "Helvetica.ttc",
-  ];
-  fonts.forEach((filename) => {
-    const link = document.createElement("link");
-    link.rel = "preload";
-    link.as = "font";
-    link.href = `_static/fonts/${filename}`;
-    link.crossOrigin = "";
-    document.head.appendChild(link);
-  });
-}
-
-const updateEditButtonLabel = () => {
-  const editButton = document.querySelector(".wy-breadcrumbs-aside a");
-  if (!editButton) return;
-  editButton.textContent = "Edit on GitHub";
-}
-
-const addHrUnderSearchForm = () => {
-  const verticalMenu = document.querySelector(".wy-menu-vertical[role=navigation]");
-  const hr = document.createElement("hr");
-  verticalMenu.parentNode.insertBefore(hr, verticalMenu);
-}
-
-const handleRstVersions = () => {
+const moveRstVersions = () => {
   const rstVersions = document.querySelector(".rst-versions");
   if (!rstVersions) return
   rstVersions.remove();
   const wyNavSide = document.querySelector("nav.wy-nav-side");
   wyNavSide.appendChild(rstVersions);
-}
-
-const rearrangeDom = () => {
-  const bodyDivs = document.querySelectorAll("body>div");
-  bodyDivs.forEach((div) => { div.remove(); });
-  const wrapperDiv = document.createElement("div");
-  wrapperDiv.classList.add(WRAPPER_CLASS);
-  bodyDivs.forEach((div) => wrapperDiv.appendChild(div));
-  document.body.prepend(wrapperDiv);
-
-  handleRstVersions()
-
-  const backdrop = document.createElement("div");
-  backdrop.classList.add("backdrop");
-  wrapperDiv.appendChild(backdrop);
-
-  const content = document.querySelector(".wy-nav-content");
-  content.id = "content";
-  const oldWrap = document.querySelector("section.wy-nav-content-wrap");
-  oldWrap.remove();
-  document.querySelector(".wy-grid-for-nav").appendChild(content);
-}
-
-const toggleMenu = (options = {}) => {
-  const handleClassToggle = ({ classList }, className) => {
-    if (typeof options.force !== "undefined") {
-      classList.toggle(className, options.force);
-    } else {
-      classList.toggle(className);
-    }
-  };
-  document
-    .querySelectorAll('[data-toggle="rst-versions"]')
-    .forEach((e) => handleClassToggle(e, MOBILE_MENU_TOGGLE_CLASS));
-  document
-    .querySelectorAll('[data-toggle="wy-nav-shift"]')
-    .forEach((e) => handleClassToggle(e, MOBILE_MENU_TOGGLE_CLASS));
-  handleClassToggle(document.querySelector(`.${WRAPPER_CLASS}`), "menu-open");
-}
-
-const buildHeader = () => {
-  const header = document.createElement("div");
-  header.classList.add("unified-header");
-  console.log(header)
-  document.querySelector(`.${WRAPPER_CLASS}`).prepend(header);
-
-  const innerHeader = document.createElement("div");
-  innerHeader.classList.add("inner-header");
-  header.appendChild(innerHeader);
-
-  const homeLink = document.createElement("a");
-  homeLink.classList.add("home-link");
-  homeLink.href = REMIX_HOME_URL;
-  homeLink.ariaLabel = "Remix project home";
-  innerHeader.appendChild(homeLink);
-
-  appendSvg(REMIX_LOGO_PATH, homeLink, REMIX_LOGO_CLASS);
-
-  const skipToContent = document.createElement("a");
-  skipToContent.classList.add("skip-to-content");
-  skipToContent.href = "#content";
-  skipToContent.innerText = "skip to content";
-  innerHeader.appendChild(skipToContent);
-
-  const navBar = document.createElement("nav");
-  navBar.classList.add("nav-bar");
-  innerHeader.appendChild(navBar);
-
-  /**
-   * type NavItem = { name: string } & ({ href: string } | { items: NavItem[] })
-   */
-  const navLinkExcludingDropdown = NAV_LINKS.filter(({ href }) => !!href)
-  const linkElements = navLinkExcludingDropdown.map(({ name, href }) => {
-    const link = document.createElement("a");
-    link.classList.add("nav-link");
-    link.setAttribute("key", name);
-    link.setAttribute("href", href);
-    link.setAttribute("aria-label", name);
-    link.innerText = name;
-    return link;
-  });
-  linkElements.forEach((link) => navBar.appendChild(link));
-
-  // Flex wrapper for color mode and mobile menu buttons
-  const navButtonContainer = document.createElement("div");
-  navButtonContainer.classList.add("nav-button-container");
-  navBar.appendChild(navButtonContainer);
-
-  // Build mobile hamburger menu
-  const menuButton = document.createElement("button");
-  menuButton.classList.add("mobile-menu-button");
-  menuButton.setAttribute("type", "button");
-  menuButton.setAttribute("aria-label", "Toggle menu");
-  menuButton.setAttribute("key", "menu button");
-  menuButton.addEventListener("click", toggleMenu);
-  appendSvg(HAMBURGER_PATH, menuButton, MOBILE_MENU_ICON_CLASS);
-  navButtonContainer.appendChild(menuButton);
 }
 
 const handleGeneralClick = (e) => {
